@@ -1,6 +1,8 @@
 source 'https://rubygems.org'
 
-gem 'rails', '3.2.12'
+ruby '2.0.0'
+
+gem 'rails', '4.0.0'
 
 gem 'rails-api'
 
@@ -9,10 +11,9 @@ gem 'airbrake'
 gem 'thin'
 gem 'decent_exposure'
 gem 'active_model_serializers'
-gem 'seedbank'
+gem 'seedbank', github: 'james2m/seedbank'
 gem 'dotenv'
-gem 'strong_parameters'
-gem 'devise'
+gem 'devise', '~> 3.1.0'
 
 group :development do
   gem 'letter_opener'
@@ -25,7 +26,7 @@ group :development do
 end
 
 group :development, :test do
-  gem 'debugger'
+  gem 'byebug', '~> 1.8.2'
   gem 'rspec-rails'
   gem 'mail_safe'
   gem 'rails_best_practices'
@@ -43,10 +44,20 @@ group :test do
 end
 
 group :development, :staging do
-  gem 'api_taster', github: 'fredwu/api_taster'
+  gem 'api_taster', github: 'fs/api_taster'
 end
 
 group :development, :test, :staging do
   gem 'faker'
   gem 'factory_girl_rails'
+end
+
+group :staging do
+  gem 'sass-rails'
+  gem 'uglifier'
+  gem 'coffee-rails'
+end
+
+group :staging, :production do
+  gem 'rails_12factor'
 end
