@@ -1,4 +1,4 @@
-Rails3BaseApi::Application.configure do
+RailsBaseApi::Application.configure do
   config.generators do |g|
     g.fixture_replacement :factory_girl, dir: 'spec/factories'
     g.stylesheets false
